Add new classes and dependencies
