+++ .vscode/settings.json
@@ -1,4 +1,7 @@
 {
   "java.configuration.updateBuildConfiguration": "automatic",
-  "cSpell.words": ["zireline"],
+  "cSpell.words": [
+    "collinearpoints",
+    "zireline"
+  ],
 }
