+++ src/main/java/com/zireline/collinearpoints/algorithm/Checker.java
@@ -1,4 +1,4 @@
-package com.zireline.algorithm;
+package com.zireline.collinearpoints.algorithm;
 
 import java.util.ArrayList;
 import java.util.List;
