+++ src/main/java/module-info.java
@@ -1,4 +1,6 @@
 module com.zireline {
-    requires javafx.controls;
-    exports com.zireline;
+  requires javafx.controls;
+  requires com.google.gson;
+
+  exports com.zireline;
 }
