+++ src/test/resources/test.json
@@ -1,0 +1,3 @@
+{
+  "testKey": "testValue"
+}
